refactor: rename socket handlers and emitters to events
