+++ src/server/events/emitters/business.socket-emitter.ts
@@ -3,7 +3,7 @@ import { emit } from '../../utils/socket.utils';
 
 // Typings
 import { PLayerBusinessEventPayloadType } from '@src/game-core';
-import { SocketEmitterArgs } from '../../typings/socket.typings';
+import { EmitterArgs } from '../../typings/events.typings';
 
 // Constants
 import { GameActions } from '../../constants';
@@ -13,7 +13,7 @@ export const emitUpdateBusinessInfo = ({
   client,
   playerInstance,
   payload
-}: SocketEmitterArgs<PLayerBusinessEventPayloadType>): void => {
+}: EmitterArgs<PLayerBusinessEventPayloadType>): void => {
   const { businessId } = payload;
   const business = playerInstance.getBusinessById(businessId);
