+++ src/server/events/emitters/game-info.socket-emitter.ts
@@ -2,7 +2,7 @@
 import { emit } from '../../utils/socket.utils';
 
 // Typings
-import { SocketEmitterBaseArgs } from '../../typings/socket.typings';
+import { EmitterBaseArgs } from '../../typings/events.typings';
 
 // Constants
 import { GameActions } from '../../constants';
@@ -11,7 +11,7 @@ import { GameActions } from '../../constants';
 export const emitUpdateUserInfo = ({
   client,
   playerInstance
-}: SocketEmitterBaseArgs): void => {
+}: EmitterBaseArgs): void => {
   const payload = { capital: playerInstance.capital };
 
   client.emit(GameActions.UPDATE_USER_INFO, { data: payload });
@@ -21,7 +21,7 @@ export const emitUpdateUserInfo = ({
 export const emitGetBusinessList = ({
   client,
   playerInstance
-}: SocketEmitterBaseArgs): void => {
+}: EmitterBaseArgs): void => {
   const businessList = playerInstance.getAllBusinessesList();
 
   emit(client, GameActions.GET_BUSINESS_LIST, { data: { businessList } });
