--- src/server/logic/index.ts
@@ -1,3 +1,0 @@
-import { startServer } from './server';
-
-export { startServer };
